Odalic project set up
This tutorial was used as an inspiration:
http://www.codingpedia.org/ama/tutorial-rest-api-design-and-implementation-in-java-with-jersey-and-spring/#2_Configuration 

For now it uses slightly older libraries, but it should not be problem
to fix this in the future. I have tested the deployment in the Tomcat,
but no thorough tests for the set up are provided.

I have provided empty implementations of the resources in the 1st
iteration API. Depending tasks are now clear to go.

This closes #7.